# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KATALA - Katalog & Reservasi Wisata Lampung API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterInputRole =
  (typeof RegisterInputRole)[keyof typeof RegisterInputRole];

export const RegisterInputRole = {
  tourist: "tourist",
  admin: "admin",
  staff: "staff",
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role?: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  tourist: "tourist",
  admin: "admin",
  staff: "staff",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Destination {
  id: number;
  name: string;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  price: number;
  dailyQuota: number;
  category: string;
  imageUrl: string;
  /** @nullable */
  rating?: number | null;
  featured?: boolean;
  createdAt?: string;
}

export interface DestinationInput {
  name: string;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  price: number;
  dailyQuota: number;
  category: string;
  imageUrl: string;
  /** @nullable */
  rating?: number | null;
  featured?: boolean;
}

export interface DestinationUpdate {
  name?: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  price?: number;
  dailyQuota?: number;
  category?: string;
  imageUrl?: string;
  /** @nullable */
  rating?: number | null;
  featured?: boolean;
}

export interface DestinationListResponse {
  data: Destination[];
  total: number;
  page: number;
  limit: number;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface Reservation {
  id: number;
  userId: number;
  destinationId: number;
  visitDate: string;
  quantity: number;
  totalPrice: number;
  status: ReservationStatus;
  destination?: Destination;
  createdAt?: string;
}

export interface ReservationInput {
  destinationId: number;
  visitDate: string;
  quantity: number;
}

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  valid: "valid",
  used: "used",
  expired: "expired",
} as const;

export interface Ticket {
  id: number;
  reservationId: number;
  ticketCode: string;
  qrCodeData: string;
  status: TicketStatus;
  /** @nullable */
  usedAt?: string | null;
  reservation?: Reservation;
  createdAt?: string;
}

export interface TicketValidateInput {
  ticketCode: string;
}

export interface TicketValidationResult {
  valid: boolean;
  message: string;
  ticket: Ticket;
}

export interface PaymentCreateInput {
  reservationId: number;
}

export interface PaymentResponse {
  snapToken: string;
  orderId: string;
  /** @nullable */
  redirectUrl?: string | null;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  expired: "expired",
  refunded: "refunded",
} as const;

export interface Transaction {
  id: number;
  reservationId: number;
  orderId: string;
  amount: number;
  status: TransactionStatus;
  /** @nullable */
  paymentType?: string | null;
  /** @nullable */
  midtransTransactionId?: string | null;
  reservation?: Reservation;
  createdAt?: string;
  updatedAt?: string;
}

export interface WebhookPayload {
  order_id: string;
  transaction_status: string;
  fraud_status: string;
  /** @nullable */
  payment_type?: string | null;
  /** @nullable */
  transaction_id?: string | null;
  /** @nullable */
  gross_amount?: string | null;
}

export interface AdminStats {
  totalDestinations: number;
  totalReservations: number;
  totalRevenue: number;
  totalUsers: number;
  pendingReservations: number;
  todayReservations: number;
}

export interface RevenuePoint {
  month: string;
  revenue: number;
  reservations: number;
}

export type ListDestinationsParams = {
  search?: string;
  category?: string;
  page?: number;
  limit?: number;
};

export type ListAllTransactionsParams = {
  status?: string;
  page?: number;
};