# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KATALA - Katalog & Reservasi Wisata Lampung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── DESTINATIONS ────────────────────────────────────────
...
  /reservations:
    get:
      operationId: listMyReservations
      tags: [reservations]
      summary: List current user's reservations
      responses:
        "200":
          description: User reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Quota exceeded or invalid
...
  /admin/transactions:
    get:
      operationId: listAllTransactions
      tags: [admin]
      summary: List all transactions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [tourist, admin, staff]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [tourist, admin, staff]
        createdAt:
          type: string

    # ─── DESTINATIONS ──────────────────────────────────────
    Destination:
      type: object
      required: [id, name, description, location, latitude, longitude, price, dailyQuota, category, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
...
    Reservation:
      type: object
      required: [id, userId, destinationId, visitDate, quantity, totalPrice, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        destinationId:
          type: integer
        visitDate:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled, expired]
        destination:
          $ref: "#/components/schemas/Destination"
        createdAt:
          type: string
    ReservationInput:
      type: object
      required: [destinationId, visitDate, quantity]
      properties:
        destinationId:
          type: integer
        visitDate:
          type: string
        quantity:
          type: integer

    # ─── TICKETS ───────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, reservationId, orderId, amount, status]
      properties:
        id:
          type: integer
        reservationId:
          type: integer
        orderId:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, failed, expired, refunded]
        paymentType:
          type: ["string", "null"]
        midtransTransactionId:
          type: ["string", "null"]
        reservation:
          $ref: "#/components/schemas/Reservation"
        createdAt:
          type: string
        updatedAt:
          type: string
...